Better share implementation between grapher and layers
