+++ python/lib/dependabot/python/dependency_grapher.rb
@@ -17,7 +17,7 @@ require "toml-rb"
 
 module Dependabot
   module Python
-    class DependencyGrapher < Dependabot::DependencyGraphers::Base # rubocop:disable Metrics/ClassLength
+    class DependencyGrapher < Dependabot::DependencyGraphers::Base
       require_relative "dependency_grapher/lockfile_generator"
       require_relative "dependency_grapher/requirements_layers"
 
@@ -140,67 +140,18 @@ module Dependabot
       # matcher recognises it as a compiled lockfile, or it is referenced (transitively) via `-r`/`-c` from a
       # retained requirements file. The last case preserves constraint/child files (e.g. `constraints.txt`) that
       # the parser needs on disk to resolve a real manifest, while still dropping bystander `.txt` files.
+      #
+      # RequirementsLayers owns the "which .txt files matter" knowledge (manifest detection and `-r`/`-c`
+      # resolution) so grouping and this bystander filter never drift apart.
       sig { void }
       def filter_non_manifest_txt_files!
-        keep = txt_files_to_keep
+        keep = RequirementsLayers.new(dependency_files: dependency_files).txt_files_to_keep
 
         file_parser.dependency_files.reject! do |file|
           next false unless file.name.end_with?(".txt")
 
           !keep.include?(file.name)
         end
-      end
-
-      # Names of the `.txt` files that must be retained: those that look like manifests or pip-compile lockfiles,
-      # plus any `.txt` reachable via `-r`/`-c` references from a retained requirements file (`.in` files are
-      # never dropped, so references originating from them are followed too).
-      sig { returns(T::Set[String]) }
-      def txt_files_to_keep
-        files_by_name = dependency_files.to_h { |file| [file.name, file] }
-
-        seeds = dependency_files.select do |file|
-          file.name.end_with?(".txt") &&
-            (RequirementsLayers.manifest_txt_filename?(file.name) ||
-              pip_compile_file_matcher.lockfile_for_pip_compile_file?(file))
-        end
-
-        # `.in` files are always retained, so a `.txt` they reference must be kept too.
-        roots = seeds + dependency_files.select { |file| file.name.end_with?(".in") }
-
-        reachable_txt_names(roots, files_by_name, Set.new(seeds.map(&:name)))
-      end
-
-      # Breadth-first closure over `-r`/`-c` references starting from `roots`, adding every referenced `.txt`
-      # file that exists in `files_by_name` to `keep`.
-      sig do
-        params(
-          roots: T::Array[Dependabot::DependencyFile],
-          files_by_name: T::Hash[String, Dependabot::DependencyFile],
-          keep: T::Set[String]
-        ).returns(T::Set[String])
-      end
-      def reachable_txt_names(roots, files_by_name, keep)
-        queue = roots.dup
-        until queue.empty?
-          file = T.must(queue.shift)
-          referenced_txt_names(file).each do |name|
-            referenced_file = files_by_name[name]
-            next if referenced_file.nil? || keep.include?(name)
-
-            keep << name
-            queue << referenced_file
-          end
-        end
-
-        keep
-      end
-
-      # Resolves the `.txt` files referenced from a requirements file via `-r`/`-c`, returning their names
-      # relative to the repo (matching the fetched DependencyFile names) so we can retain them. Delegates to
-      # RequirementsLayers.referenced_paths so grouping and the bystander filter resolve references identically.
-      sig { params(file: Dependabot::DependencyFile).returns(T::Array[String]) }
-      def referenced_txt_names(file)
-        RequirementsLayers.referenced_paths(file).select { |path| path.end_with?(".txt") }
       end
 
       sig { returns(T::Boolean) }
