+++ python/lib/dependabot/python/dependency_grapher/requirements_layers.rb
@@ -6,6 +6,7 @@ require "sorbet-runtime"
 
 require "dependabot/dependency_file"
 require "dependabot/dependency_graphers/base"
+require "dependabot/python/pip_compile_file_matcher"
 require "dependabot/python/shared_file_fetcher"
 
 module Dependabot
@@ -97,10 +98,67 @@ module Dependabot
           end
         end
 
+        # Names of the `.txt` files that must be retained: those that look like manifests or pip-compile
+        # lockfiles, plus any `.txt` reachable via `-r`/`-c` references from a retained requirements file (`.in`
+        # files are never dropped, so references originating from them are followed too).
+        #
+        # Shared with the grapher's bystander filter so the same knowledge that groups layers also decides which
+        # `.txt` files survive, keeping the two in lockstep.
+        sig { returns(T::Set[String]) }
+        def txt_files_to_keep
+          files_by_name = dependency_files.to_h { |file| [file.name, file] }
+
+          seeds = dependency_files.select do |file|
+            file.name.end_with?(".txt") &&
+              (self.class.manifest_txt_filename?(file.name) ||
+                pip_compile_file_matcher.lockfile_for_pip_compile_file?(file))
+          end
+
+          # `.in` files are always retained, so a `.txt` they reference must be kept too.
+          roots = seeds + dependency_files.select { |file| file.name.end_with?(".in") }
+
+          reachable_txt_names(roots, files_by_name, Set.new(seeds.map(&:name)))
+        end
+
         private
 
         sig { returns(T::Array[Dependabot::DependencyFile]) }
         attr_reader :dependency_files
+
+        # Breadth-first closure over `-r`/`-c` references starting from `roots`, adding every referenced `.txt`
+        # file that exists in `files_by_name` to `keep`.
+        sig do
+          params(
+            roots: T::Array[Dependabot::DependencyFile],
+            files_by_name: T::Hash[String, Dependabot::DependencyFile],
+            keep: T::Set[String]
+          ).returns(T::Set[String])
+        end
+        def reachable_txt_names(roots, files_by_name, keep)
+          queue = roots.dup
+          until queue.empty?
+            file = T.must(queue.shift)
+            self.class.referenced_paths(file).each do |name|
+              next unless name.end_with?(".txt")
+
+              referenced_file = files_by_name[name]
+              next if referenced_file.nil? || keep.include?(name)
+
+              keep << name
+              queue << referenced_file
+            end
+          end
+
+          keep
+        end
+
+        sig { returns(PipCompileFileMatcher) }
+        def pip_compile_file_matcher
+          @pip_compile_file_matcher ||= T.let(
+            PipCompileFileMatcher.new(dependency_files.select { |f| f.name.end_with?(".in") }),
+            T.nilable(PipCompileFileMatcher)
+          )
+        end
 
         # The set of files that act as the owning manifest for a layer: every requirements `.txt` that looks
         # like a real manifest, plus any `.in` file that has no compiled `.txt` counterpart.
